+++ deepsort/generate_detections.py
@@ -51,7 +51,7 @@ def extract_image_patch(image, bbox, patch_shape):
 
     # convert to top left, bottom right
     bbox[2:] += bbox[:2]
-    bbox = bbox.astype(np.int)
+    bbox = bbox.astype(np.int32)
 
     # clip at image boundaries
     bbox[:2] = np.maximum(0, bbox[:2])
